requirements.txt
livekit-agents==0.8.12
livekit-plugins-openai==0.8.5
livekit-plugins-silero==0.7.6
python-dotenv>=1.0.0
pypdf
